Sprint 5 (#591 E5a): extract SubmissionLifecycleService from SubmissionHandler
Critical-path decomposition done conservatively — SubmissionHandler's public API
stays 100% intact (all 24 callers + tests unchanged). The lifecycle/maintenance
operations (trash/restore/delete, bulk_trash/restore/delete,
move_submissions_between_forms, delete_all_submissions, reset_migration_counters,
reset_submission_counter, run_data_cleanup) move to a new
SubmissionLifecycleService; SubmissionHandler keeps a thin same-signature
delegator for each. The create flow (process_submission) and the read/update/
token methods are untouched.

The service reads the repository via the handler at call-time (get_repository()),
so the test's post-construction mock swap is honored. Repository calls, cache
invalidation, ActivityLog disable/enable ordering, WP actions, return shapes and
guards are byte-identical. SubmissionHandler 871 → 624 lines.

Service carries @covers + setUp preload (verified 60/133). Behavior-preserving —
Submission suite green (352 tests); PHPStan L8 + WPCS clean.
