+++ tests/Unit/SubmissionHandlerTest.php
@@ -20,6 +20,7 @@ use FreeFormCertificate\Core\Encryption;
  * is always active.
  * @runClassInSeparateProcess
  * @preserveGlobalState disabled
+ * @covers \FreeFormCertificate\Submissions\SubmissionLifecycleService
  */
 class SubmissionHandlerTest extends TestCase {
 
@@ -31,6 +32,11 @@ class SubmissionHandlerTest extends TestCase {
     protected function setUp(): void {
         parent::setUp();
         Monkey\setUp();
+
+        // pcov does not record lines for files first autoloaded mid-test-method,
+        // so the lifecycle service's coverage would attribute to nothing. Preload
+        // the extracted class here so pcov attributes its lines to this test.
+        class_exists( '\\FreeFormCertificate\\Submissions\\SubmissionLifecycleService' );
 
         // Mock global $wpdb
         global $wpdb;
